syncing limited to blur
